Add --existing-mnemonic opt to deposit cli wrapper, refactor

+++ src/run-staking-deposit-cli.sh
@@ -4,78 +4,95 @@
 
 this_dir="$(realpath "$(dirname "${BASH_SOURCE[0]}")")"
 source "$this_dir/common.sh"
+source "$this_dir/staking_deposit_cli.sh"
 housekeeping
+
+function _show_usage() {
+	cat >&2 <<-EOF 
+	usage:
+	create-validator-keys.sh [-h|--help] {--new-mnemonic|--existing-mnemonic}
+	EOF
+}
+
+_mode_new=false
+_mode_existing=false
+
+function _parse_arg() {
+	if [[ $1 == '-h' || $1 == '--help' ]]; then
+		_show_usage
+		exit 0
+	elif [[ $1 == '--new-mnemonic' ]]; then
+		_mode_new=true
+	elif [[ $1 == '--existing-mnemonic' ]]; then
+		_mode_existing=true
+	else
+		printerr "unknown argument: $1"
+		_show_usage
+		exit 1
+	fi
+}
+
+function _validate() {
+	[[ $_mode_new == true || $_mode_existing == true ]]
+}
+
+if [[ $# -gt 0 ]]; then
+	_parse_arg "$1"
+	shift
+	if ! _validate && [[ $# -gt 0 ]]; then
+		_parse_arg "$1"
+		shift
+	fi
+fi
+
+if ! _validate; then
+	_show_usage
+	exit 1
+fi
 
 # -------------------------- BANNER -------------------------------------------
 
 echo -n "${color_blue}${bold}"
 cat <<EOF
-██████╗ ██╗   ██╗███╗   ██╗                                                
-██╔══██╗██║   ██║████╗  ██║                                                
-██████╔╝██║   ██║██╔██╗ ██║█████╗                                          
-██╔══██╗██║   ██║██║╚██╗██║╚════╝                                          
-██║  ██║╚██████╔╝██║ ╚████║                                                
-╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═══╝                                                
-                                                                           
-██████╗ ███████╗██████╗  ██████╗ ███████╗██╗████████╗    ██████╗██╗     ██╗
-██╔══██╗██╔════╝██╔══██╗██╔═══██╗██╔════╝██║╚══██╔══╝   ██╔════╝██║     ██║
-██║  ██║█████╗  ██████╔╝██║   ██║███████╗██║   ██║█████╗██║     ██║     ██║
-██║  ██║██╔══╝  ██╔═══╝ ██║   ██║╚════██║██║   ██║╚════╝██║     ██║     ██║
-██████╔╝███████╗██║     ╚██████╔╝███████║██║   ██║      ╚██████╗███████╗██║
-╚═════╝ ╚══════╝╚═╝      ╚═════╝ ╚══════╝╚═╝   ╚═╝       ╚═════╝╚══════╝╚═╝
+░█▀▀░█▀▄░█▀▀░█▀█░▀█▀░█▀▀░░░░░█▀█░█▀▀░█░█░░░░░█░█░█▀▀░█░█░█▀▀
+░█░░░█▀▄░█▀▀░█▀█░░█░░█▀▀░▄▄▄░█░█░█▀▀░█▄█░▄▄▄░█▀▄░█▀▀░░█░░▀▀█
+░▀▀▀░▀░▀░▀▀▀░▀░▀░░▀░░▀▀▀░░░░░▀░▀░▀▀▀░▀░▀░░░░░▀░▀░▀▀▀░░▀░░▀▀▀
 ${color_reset}
 EOF
 
 # -------------------------- PREAMBLE -----------------------------------------
 
+preamble="[${theme_value}New Mnemonic${color_reset}] Generates a new mnemonic and creates validator keys with it on the air-gapped PC."
+if [[ $_mode_existing == true ]]; then
+	preamble="[${theme_value}Existing Mnemonic${color_reset}] Uses an existing mnemonic and creates validator keys with it on the air-gapped PC."
+fi
+
 cat <<EOF
-Runs the Ethereum Staking Deposit CLI on the air-gapped PC.
+$preamble
+See: https://github.com/ethereum/staking-deposit-cli?tab=readme-ov-file#step-2-create-keys-and-deposit_data-json
 EOF
 press_any_key_to_continue
 
 # -------------------------- PRECONDITIONS ------------------------------------
 
-assert_offline
-assert_not_on_node_server
-assert_sudo
-
-check_is_defined ethereum_staking_deposit_cli_version
-check_is_defined ethereum_staking_deposit_cli_sha256_checksum
-check_is_defined ethereum_staking_deposit_cli_url
-
-deposit_cli_basename="$(basename "$ethereum_staking_deposit_cli_url")"
-deposit_cli_basename_sha256="${deposit_cli_basename}.sha256"
-deposit_cli="$this_dir/$deposit_cli_basename"
-deposit_cli_sha256="$this_dir/$deposit_cli_basename_sha256"
-
-check_file_exists --sudo deposit_cli
-check_file_exists --sudo deposit_cli_sha256
-
-print_failed_checks --error || exit 1
+staking_deposit_cli__preconditions || exit
 
 validator_keys_parent_dir="$this_dir"
 validator_keys_dir="$validator_keys_parent_dir/validator_keys"
 
-# if validator_keys directory already exists, confirm overwrite
-reset_checks
-check_directory_does_not_exist --sudo validator_keys_dir
-if ! print_failed_checks --warn; then
-	continue_or_exit 1 "Overwrite?"
-	printf '\n'
+if [[ $_mode_new == true ]]; then
+	# if validator_keys directory already exists, confirm overwrite
+	reset_checks
+	check_directory_does_not_exist --sudo validator_keys_dir
+	if ! print_failed_checks --warn; then
+		continue_or_exit 1 "Overwrite?"
+		printf '\n'
+	fi
 fi
 
 # -------------------------- RECONNAISSANCE -----------------------------------
 
-printf '%s' \
-	"Using an online PC, navigate to " \
-	"${color_blue}https://github.com/ethereum/staking-deposit-cli/releases/tag/${ethereum_staking_deposit_cli_version}${color_reset} " \
-	"and verify the ${color_lightgray}SHA256 Checksum${color_reset} of ${theme_filename}$deposit_cli_basename${color_reset}"
-printf '\n'
-if ! yes_or_no --default-no "Does it match this? ${theme_value}$ethereum_staking_deposit_cli_sha256_checksum${color_reset}"; then
-	printerr "unexpected checksum; ensure that ${color_lightgray}ethereum_staking_deposit_cli_${color_reset} values in ${theme_filename}env.sh${color_reset} are correct and relaunch this script"
-	exit 1
-fi
-printf '\n'
+staking_deposit_cli__reconnaissance || exit
 
 read_default "Number of validator keys to generate" 1 num_validators
 if [[ ! $num_validators =~ ^[[:digit:]]+$ || ! $num_validators -gt 0 ]]; then
@@ -83,6 +100,15 @@ if [[ ! $num_validators =~ ^[[:digit:]]+$ || ! $num_validators -gt 0 ]]; then
 	exit 1
 fi
 printf '\n'
+
+if [[ $_mode_existing == true ]]; then
+	read_default "Validator start index (0-based)" 1 validator_start_index
+	if [[ ! $validator_start_index =~ ^[[:digit:]]+$ || ! $validator_start_index -ge 0 ]]; then
+		printerr "must choose an integer ≥ 0"
+		exit 1
+	fi
+	printf '\n'
+fi
 
 # -------------------------- EXECUTION ----------------------------------------
 
@@ -100,54 +126,50 @@ trap 'on_err_noretry' ERR
 trap 'on_exit' EXIT
 
 assert_sudo
+staking_deposit_cli__unpack_tarball
 
-# # change to the dist directory and chown it
-# cd "$this_dir"
-# copy to the temp dir and chown it all
-cp -f "$deposit_cli" ./
-cp -f "$deposit_cli_sha256" ./
-sudo chown -R "$USER:$USER" ./
+if [[ $_mode_new == true ]]; then
+	# confirmation message
+	cat <<-EOF
+	Ready to run the following command:${color_lightgray}
+	$deposit_cli_bin --language=English new-mnemonic \\
+		--num_validators=$num_validators \\
+		--mnemonic_language=English \\
+		--chain="$ethereum_network" \\
+		--folder="$validator_keys_parent_dir"
+	${color_reset}
+	EOF
+	continue_or_exit 1
 
-# checksum using the included .sha256 file
-printinfo "Verifying SHA256 checksum..."
-sha256sum -c "$deposit_cli_sha256" # trapped
-printf '\n'
+	# remove any existing keystore
+	rm -rfv "$validator_keys_dir" &>/dev/null
 
-# unpack tarball
-printinfo "Unpacking tarball..."
-tar xvzf "$deposit_cli_basename"
-printf '\n'
+	# generate the key(s)
+	$deposit_cli_bin --language=English new-mnemonic \
+		--num_validators=$num_validators \
+		--mnemonic_language=English \
+		--chain="$ethereum_network" \
+		--folder="$validator_keys_parent_dir"
+else
+	# confirmation message
+	cat <<-EOF
+	Ready to run the following command:${color_lightgray}
+	$deposit_cli_bin --language=English existing-mnemonic \\
+		--validator_start_index=$validator_start_index \\
+		--num_validators=$num_validators \\
+		--chain="$ethereum_network" \\
+		--folder="$validator_keys_parent_dir"
+	${color_reset}
+	EOF
+	continue_or_exit 1
 
-deposit_cli_dir="${deposit_cli_basename%%.*}"
-deposit_cli_bin="$deposit_cli_dir/deposit"
-
-# assert that all is well so far
-reset_checks
-check_directory_exists --sudo deposit_cli_dir
-check_executable_exists --sudo deposit_cli_bin
-print_failed_checks --error # trapped
-
-# confirmation message
-cat <<EOF
-Ready to run the following command:${color_lightgray}
-$deposit_cli_bin --language=English new-mnemonic \\
-	--num_validators=$num_validators \\
-	--mnemonic_language=English \\
-	--chain="$ethereum_network" \\
-	--folder="$validator_keys_parent_dir"
-${color_reset}
-EOF
-continue_or_exit 1
-
-# remove any existing keystore
-rm -rfv "$validator_keys_dir" &>/dev/null
-
-# generate the key(s)
-$deposit_cli_bin --language=English new-mnemonic \
-	--num_validators=$num_validators \
-	--mnemonic_language=English \
-	--chain="$ethereum_network" \
-	--folder="$validator_keys_parent_dir"
+	# generate the key(s)
+	$deposit_cli_bin --language=English existing-mnemonic \
+		--validator_start_index=$validator_start_index \
+		--num_validators=$num_validators \
+		--chain="$ethereum_network" \
+		--folder="$validator_keys_parent_dir"
+fi
 
 # -------------------------- POSTCONDITIONS -----------------------------------
